w2w_api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
